requirements.txt
easyocr==1.4.2
numpy==1.24.3
opencv-python-headless==4.8.1.78
pillow==10.1.0
PyMuPDF==1.23.8